Eliminate the unused legacy GPX implementation completely.
