+++ src/chart1.cpp
@@ -1537,6 +1537,8 @@ bool MyApp::OnInit()
 {
     if( !wxApp::OnInit() ) return false;
 
+    GpxDocument::SeedRandom();
+    
     last_own_ship_sog_cog_calc_ts = wxInvalidDateTime;
 
 #if defined(__WXGTK__)  && defined(ocpnUSE_GLES) && defined(__ARM_ARCH)
@@ -3574,7 +3576,7 @@ void MyFrame::OnCloseWindow( wxCloseEvent& event )
 
             wxString name = now.Format();
             name.Prepend( _("Anchorage created ") );
-            RoutePoint *pWP = new RoutePoint( gLat, gLon, _T("anchorage"), name, GPX_EMPTY_STRING );
+            RoutePoint *pWP = new RoutePoint( gLat, gLon, _T("anchorage"), name, wxEmptyString );
             pWP->m_bShowName = false;
             pWP->m_bIsolatedMark = true;
 
@@ -4735,7 +4737,7 @@ void MyFrame::ActivateMOB( void )
     mob_label += _T(" at ");
     mob_label += mob_time.FormatTime();
 
-    RoutePoint *pWP_MOB = new RoutePoint( gLat, gLon, _T ( "mob" ), mob_label, GPX_EMPTY_STRING );
+    RoutePoint *pWP_MOB = new RoutePoint( gLat, gLon, _T ( "mob" ), mob_label, wxEmptyString );
     pWP_MOB->m_bKeepXRoute = true;
     pWP_MOB->m_bIsolatedMark = true;
     pWP_MOB->SetWaypointArrivalRadius( -1.0 ); // Negative distance is code to signal "Never Arrive"
@@ -4750,7 +4752,7 @@ void MyFrame::ActivateMOB( void )
         ll_gc_ll( gLat, gLon, gCog, 1.0, &zlat, &zlon );
 
         RoutePoint *pWP_src = new RoutePoint( zlat, zlon, g_default_wp_icon,
-                wxString( _( "1.0 NM along COG" ) ), GPX_EMPTY_STRING );
+                wxString( _( "1.0 NM along COG" ) ), wxEmptyString );
         pSelect->AddSelectableRoutePoint( zlat, zlon, pWP_src );
 
         Route *temp_route = new Route();
@@ -9918,7 +9920,7 @@ void MyFrame::ActivateAISMOBRoute( AIS_Target_Data *ptarget )
     mob_label += _T(" at ");
     mob_label += mob_time.FormatTime();
 
-    RoutePoint *pWP_MOB = new RoutePoint( ptarget->Lat, ptarget->Lon, _T ( "mob" ), mob_label, GPX_EMPTY_STRING );
+    RoutePoint *pWP_MOB = new RoutePoint( ptarget->Lat, ptarget->Lon, _T ( "mob" ), mob_label, wxEmptyString );
     pWP_MOB->m_bKeepXRoute = true;
     pWP_MOB->m_bIsolatedMark = true;
     pSelect->AddSelectableRoutePoint( ptarget->Lat, ptarget->Lon, pWP_MOB );
@@ -9928,7 +9930,7 @@ void MyFrame::ActivateAISMOBRoute( AIS_Target_Data *ptarget )
     /* We want to start tracking any MOB in range (Which will trigger false alarms with messages received over the network etc., but will a) not discard nearby event even in case our GPS is momentarily unavailable and b) work even when the boat is stationary, in which case some GPS units do not provide COG)
     if( bGPSValid && !wxIsNaN(gCog) && !wxIsNaN(gSog) ) { */
         RoutePoint *pWP_src = new RoutePoint( gLat, gLon, g_default_wp_icon,
-                                              wxString( _( "Own ship" ) ), GPX_EMPTY_STRING );
+                                              wxString( _( "Own ship" ) ), wxEmptyString );
         pSelect->AddSelectableRoutePoint( gLat, gLon, pWP_src );
 
         pAISMOBRoute = new Route();
