+++ src/navutil.cpp
@@ -57,8 +57,6 @@
 #include "routeprop.h"
 #include "s52utils.h"
 #include "chartbase.h"
-#include "tinyxml.h"
-#include "gpxdocument.h"
 #include "ocpndc.h"
 #include "geodesic.h"
 #include "datastream.h"
@@ -4222,6 +4220,59 @@ void AlphaBlending( ocpnDC &dc, int x, int y, int size_x, int size_y, float radi
     }
 }
 
+// RFC4122 version 4 compliant random UUIDs generator.
+wxString GpxDocument::GetUUID(void)
+{
+    wxString str;
+    struct {
+        int time_low;
+        int time_mid;
+        int time_hi_and_version;
+        int clock_seq_hi_and_rsv;
+        int clock_seq_low;
+        int node_hi;
+        int node_low;
+    } uuid;
+    
+    uuid.time_low = GetRandomNumber(0, 2147483647);//FIXME: the max should be set to something like MAXINT32, but it doesn't compile un gcc...
+    uuid.time_mid = GetRandomNumber(0, 65535);
+    uuid.time_hi_and_version = GetRandomNumber(0, 65535);
+    uuid.clock_seq_hi_and_rsv = GetRandomNumber(0, 255);
+    uuid.clock_seq_low = GetRandomNumber(0, 255);
+    uuid.node_hi = GetRandomNumber(0, 65535);
+    uuid.node_low = GetRandomNumber(0, 2147483647);
+    
+    /* Set the two most significant bits (bits 6 and 7) of the
+     * clock_seq_hi_and_rsv to zero and one, respectively. */
+    uuid.clock_seq_hi_and_rsv = (uuid.clock_seq_hi_and_rsv & 0x3F) | 0x80;
+    
+    /* Set the four most significant bits (bits 12 through 15) of the
+     * time_hi_and_version field to 4 */
+    uuid.time_hi_and_version = (uuid.time_hi_and_version & 0x0fff) | 0x4000;
+    
+    str.Printf(_T("%08x-%04x-%04x-%02x%02x-%04x%08x"),
+               uuid.time_low,
+               uuid.time_mid,
+               uuid.time_hi_and_version,
+               uuid.clock_seq_hi_and_rsv,
+               uuid.clock_seq_low,
+               uuid.node_hi,
+               uuid.node_low);
+    
+    return str;
+}
 
+int GpxDocument::GetRandomNumber(int range_min, int range_max)
+{
+    long u = (long)wxRound(((double)rand() / ((double)(RAND_MAX) + 1) * (range_max - range_min)) + range_min);
+    return (int)u;
+}
 
-
+void GpxDocument::SeedRandom()
+{
+    /* Fill with random. Miliseconds hopefully good enough for our usage, reading /dev/random would be much better on linux and system guid function on Windows as well */
+    wxDateTime x = wxDateTime::UNow();
+    long seed = x.GetMillisecond();
+    seed *= x.GetTicks();
+    srand(seed);
+}
